appapi/requirements.txt
###### Requirements without Version Specifiers ######`
datasets
###### Requirements with Version Specifiers ######`
fastapi==0.78.0 
uvicorn==0.17.6 
elasticsearch==8.4.3
spacy==3.4.1
ja-ginza-electra==5.1.2
ipadic==1.0.0
fugashi[unidic-lite]==1.2.0
sentencepiece==0.1.97
torch==1.12.1
simpletransformers==0.63.9
pyknp==0.6.1
fastapi==0.78.0 
uvicorn==0.17.6 